Alterei os tipos de id para Long

+++ cinemaAutoatendimento/src/main/java/com/example/cinemaAutoatendimento/controller/AssentoController.java
@@ -22,35 +22,35 @@ public class AssentoController {
     private PessoaService pessoaService;
 
     @PostMapping("/sessao/{id}")
-    AssentoModel salvarAssento(@PathVariable int id, @RequestBody AssentoModel assento){
+    AssentoModel salvarAssento(@PathVariable Long id, @RequestBody AssentoModel assento){
         return assentoService.salvarAssento(id, assento);
     }
 
     @PutMapping("/ocupar/{id}")
-    public AssentoModel ocuparAssento(@PathVariable int id, @RequestParam int pessoaId){
+    public AssentoModel ocuparAssento(@PathVariable Long id, @RequestParam Long pessoaId){
         PessoaModel pessoa = pessoaService.buscarPessoaPorId(pessoaId);
         return assentoService.ocuparAssento(id, pessoa);
     }
 
     @PutMapping("/desocupar/{id}")
-    public AssentoModel desocuparAssento(@PathVariable int id){
+    public AssentoModel desocuparAssento(@PathVariable Long id){
         return assentoService.desocuparAssento(id);
     }
 
     //vou colocar só por sessão, não tem sentido algum listar assentos aleatórios de todas as sessões...
     @GetMapping("/sessao/{sessaoId}")
-    public List<AssentoModel> listarAssentosPorSessao(@PathVariable int sessaoId){
+    public List<AssentoModel> listarAssentosPorSessao(@PathVariable Long sessaoId){
         return assentoService.listarAssentosPorSessao(sessaoId);
     }
 
     //busca apenas os livres por sessão
     @GetMapping("/livres/sessao/{sessaoId}")
-    public List<AssentoModel> listarAssentosLivresPorSessao(@PathVariable int sessaoId){
+    public List<AssentoModel> listarAssentosLivresPorSessao(@PathVariable Long sessaoId){
         return assentoService.listarAssentosLivresPorSessao(sessaoId);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> excluirAssento(@PathVariable int id){
+    public ResponseEntity<Void> excluirAssento(@PathVariable Long id){
         boolean deleted = assentoService.excluirAssento(id);
         if (deleted){
             return ResponseEntity.noContent().build();
